requirements.txt
Django==3.1.7
requests==2.24.0
torch==1.7.1
numpy==1.19.5
torchvision==0.8.2
Pillow==8.1.2
Django==3.1.7
requests==2.24.0
